add alert and show alert if shipment is closed

+++ client/src/app/admin/all-shipment/all-shipment.component.ts
@@ -8,6 +8,7 @@ import {AssignWorkersComponent} from '../assign-workers/assign-workers.component
 import {FormControl} from '@angular/forms';
 import {ShipmentService} from '../../shared/shipment.service';
 import {UserService} from '../../shared/user.service';
+import Swal from 'sweetalert2';
 
 export interface ShipmentSchema {
   id: string;
@@ -96,10 +97,27 @@ export class AllShipmentComponent implements OnInit {
   }
 
   deleteShipmentHandler(id) {
-    this.shipmentService.deleteShipment(id)
-      .subscribe(res => {
-        this.getAllShipments();
-      });
+    Swal.fire({
+      title: 'Are you sure?',
+      text: 'You won\'t be able to revert this!',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.value) {
+        this.shipmentService.deleteShipment(id)
+          .subscribe(res => {
+            Swal.fire(
+              'Deleted!',
+              'Your data has been deleted.',
+              'success'
+            );
+            this.getAllShipments();
+          });
+      }
+    });
   }
 
 }
